MilicaCosminSergiu_licenta/MilicaCosminSergiu_licenta/PlatformaWeb/Cache_Service/requirements.txt
fastapi~=0.104.1
uvicorn
peewee~=3.17.0
pymongo
requests
PyJWT
grpcio
protobuf
redis
boto3